requirements.txt
appdirs==1.4.3
aspy.yaml==1.3.0
attrs==19.1.0
black==19.3b0
certifi==2019.6.16
cfgv==2.0.0
chardet==3.0.4
click==7.0
identify==1.4.5
idna==2.8
importlib-metadata==0.18
nodeenv==1.3.3
pre-commit==1.17.0
pytest-bdd==3.1.1
pytest-mock==1.10.4
pytest==5.0.1
pyyaml==5.1.1
requests==2.22.0
six==1.12.0
toml==0.10.0
urllib3==1.25.3
virtualenv==16.6.2
zipp==0.5.2
